Fix wrong asynchronous pipe handling
When data were received asynchronously and there was issued splice(2) with
AF_KTLS as source, sequence number was not correctly incremented and the record
was not popped from TCP/UDP socket. This caused correct decryption every time
because of same record on TCP/UDP socket, but socket returned same data every
call.